added item/actor definition system
items/actors/gObjects now have a definition which is supposed to keep
the base, non-changing-between-instances info about the item. might be a
bit clumpy atm, but hey, it works..? nws, means that we can save all
defs to file, load in at the beginning of the game &c, noice. also made
actors automatically generate a corpse item definition.