basic component system
basic component system to counter the Item-class that is sort of
starting to get a bit too large for my taste. conveniently enough, we
can then also easily tell whether an item is wearable or not by simply
checking if it has a cWearable. not yet phased out the old variables,
that's the next step.