containers w/ items i.s.o. IDs, basic keybinding &c
containers now actually hold items instead of item IDs, since it was
annoying having them separate from the way the actor inventories work,
and I think that the containersaving/loading is going to be robust
enough this way. added a bit of interacting with stuff in the inventory
screen as well (wielding, wearing &c), and made a small basic keybinding
system (so we in the future don't need to hunt down keys in the source
if we want to change one of them...). generate number of hands from item
weight.